requirements.txt
asgiref==3.11.0
Django==5.2.9
django-cors-headers==4.9.0
django-import-export==4.3.14
djangorestframework==3.16.1
oracledb==3.4.1
mysql-connector-python==8.0.33
mysqlclient==2.2.7
psycopg2-binary==2.9.11
Pillow==12.0.0
sqlparse==0.5.4
requests==2.31.0
pandas==2.0.3
numpy==1.25.2
tablib==3.9.0
pytz==2023.3
python-dateutil==2.8.2
PyYAML==6.0.1
pypdf==6.4.0
tablib[html]==3.9.0
tablib[xlsx]==3.9.0
tablib[yaml]==3.9.0
coreapi==2.3.3
coreschema==0.0.4
markdown==3.7
django-filter==24.3
psycopg[binary]==3.2.4
cryptography==46.0.3
python-dotenv==1.0.1
